+++ drizzle.config.ts
@@ -1,12 +1,13 @@
-import { type Config } from "drizzle-kit";
+import { defineConfig } from "drizzle-kit";
 
 import { env } from "@/env";
 
-export default {
-  schema: "./src/server/db/schema/index.ts",
+export default defineConfig({
+  out: "./drizzle",
   dialect: "postgresql",
+  schema: "./src/server/db/schema/index.ts",
   dbCredentials: {
     url: env.DATABASE_URL,
   },
-  // tablesFilter: ["bcc3_*"],
-} satisfies Config;
+  casing: "snake_case",
+});