+++ src/server/db/index.ts
@@ -1,21 +1,8 @@
-import { drizzle } from "drizzle-orm/postgres-js";
-import postgres from "postgres";
-
 import { env } from "@/env";
+import { drizzle } from "drizzle-orm/postgres-js";
 import * as schema from "./schema/index";
 
-/**
- * Cache the database connection in development. This avoids creating a new connection on every HMR
- * update.
- */
-const globalForDb = globalThis as unknown as {
-  conn: postgres.Sql | undefined;
-};
-
-export const conn = globalForDb.conn ?? postgres(env.DATABASE_URL, { prepare: false });
-if (env.NODE_ENV !== "production") globalForDb.conn = conn;
-
-export const db = drizzle(conn, { schema });
+export const db = drizzle(env.DATABASE_URL, { schema, casing: "snake_case" });
 
 export type db = typeof db;