+++ src/server/db/schema/repeatingRide.ts
@@ -1,33 +1,33 @@
 import { relations } from "drizzle-orm";
-import { index, integer, text, timestamp, varchar } from "drizzle-orm/pg-core";
+import * as t from "drizzle-orm/pg-core";
 import { createTable } from "./create-table";
 import rides from "./ride";
 
 const repeatingRides = createTable(
   "repeating_rides",
   {
-    id: text("id")
+    id: t
+      .text()
       .primaryKey()
       .$defaultFn(() => crypto.randomUUID()),
-    name: varchar("name", { length: 255 }).notNull(),
-    schedule: text("schedule").notNull(),
-    winterStartTime: varchar("winter_start_time", { length: 255 }),
-    rideGroup: varchar("ride_group", { length: 255 }),
-    destination: varchar("destination", { length: 255 }),
-    distance: integer("distance"),
-    meetPoint: varchar("meet_point", { length: 255 }),
-    route: varchar("route", { length: 255 }),
-    leader: varchar("leader", { length: 255 }),
-    notes: text("notes"),
-    rideLimit: integer("ride_limit").default(-1).notNull(),
-    createdAt: timestamp("created_at", { precision: 3, mode: "string" })
+    name: t.varchar({ length: 255 }).notNull(),
+    schedule: t.text().notNull(),
+    winterStartTime: t.varchar({ length: 255 }),
+    rideGroup: t.varchar({ length: 255 }),
+    destination: t.varchar({ length: 255 }),
+    distance: t.integer(),
+    meetPoint: t.varchar({ length: 255 }),
+    route: t.varchar({ length: 255 }),
+    leader: t.varchar({ length: 255 }),
+    notes: t.text(),
+    rideLimit: t.integer().default(-1).notNull(),
+    createdAt: t
+      .timestamp({ precision: 3, mode: "string" })
       .defaultNow()
       .notNull(),
-    updatedAt: timestamp("updated_at").notNull().defaultNow(),
+    updatedAt: t.timestamp().notNull().defaultNow(),
   },
-  (ride) => ({
-    repeatingRideIndex: index().on(ride.name),
-  }),
+  (table) => [t.index().on(table.name)],
 );
 
 export default repeatingRides;