+++ src/server/db/schema/ride.ts
@@ -1,13 +1,5 @@
 import { relations } from "drizzle-orm";
-import {
-  boolean,
-  index,
-  integer,
-  text,
-  timestamp,
-  varchar,
-} from "drizzle-orm/pg-core";
-
+import * as t from "drizzle-orm/pg-core";
 import { createTable } from "./create-table";
 import repeatingRides from "./repeatingRide";
 import userOnRides from "./usersOnRide";
@@ -15,35 +7,38 @@ import userOnRides from "./usersOnRide";
 const rides = createTable(
   "rides",
   {
-    id: text("id")
+    id: t
+      .text()
       .primaryKey()
       .$defaultFn(() => crypto.randomUUID()),
-    name: varchar("name", { length: 255 }).notNull(),
-    rideGroup: varchar("ride_group", { length: 255 }),
-    rideDate: timestamp("ride_date", {
-      precision: 3,
-      mode: "string",
-    }).notNull(),
-    destination: varchar("destination", { length: 255 }),
-    distance: integer("distance"),
-    meetPoint: varchar("meet_point", { length: 255 }),
-    route: varchar("route", { length: 255 }),
-    leader: varchar("leader", { length: 255 }),
-    notes: text("notes"),
-    rideLimit: integer("ride_limit").notNull().default(-1),
-    deleted: boolean("deleted").notNull().default(false),
-    cancelled: boolean("cancelled").notNull().default(false),
-    scheduleId: text("schedule_id"),
-    createdAt: timestamp("created_at", { precision: 3, mode: "string" })
+    name: t.varchar({ length: 255 }).notNull(),
+    rideGroup: t.varchar({ length: 255 }),
+    rideDate: t
+      .timestamp({
+        precision: 3,
+        mode: "string",
+      })
+      .notNull(),
+    destination: t.varchar({ length: 255 }),
+    distance: t.integer(),
+    meetPoint: t.varchar({ length: 255 }),
+    route: t.varchar({ length: 255 }),
+    leader: t.varchar({ length: 255 }),
+    notes: t.text(),
+    rideLimit: t.integer().notNull().default(-1),
+    deleted: t.boolean().notNull().default(false),
+    cancelled: t.boolean().notNull().default(false),
+    scheduleId: t.text(),
+    createdAt: t
+      .timestamp({ precision: 3, mode: "string" })
       .defaultNow()
       .notNull(),
-    updatedAt: timestamp("updated_at", { precision: 3, mode: "string" })
+    updatedAt: t
+      .timestamp({ precision: 3, mode: "string" })
       .defaultNow()
       .notNull(),
   },
-  (ride) => ({
-    rideIndex: index().on(ride.name),
-  }),
+  (table) => [t.index().on(table.name)],
 );
 
 export const rideRelations = relations(rides, ({ one, many }) => ({