+++ src/server/db/schema/session.ts
@@ -1,25 +1,24 @@
 import { relations } from "drizzle-orm";
-import { index, timestamp, varchar } from "drizzle-orm/pg-core";
+import * as t from "drizzle-orm/pg-core";
 import { createTable } from "./create-table";
 import users from "./user";
 
 const sessions = createTable(
   "sessions",
   {
-    sessionToken: varchar("session_token", { length: 255 })
-      .notNull()
-      .primaryKey(),
-    userId: varchar("user_id", { length: 255 })
+    sessionToken: t.varchar({ length: 255 }).notNull().primaryKey(),
+    userId: t
+      .varchar({ length: 255 })
       .notNull()
       .references(() => users.id),
-    expires: timestamp("expires", {
-      mode: "date",
-      withTimezone: true,
-    }).notNull(),
+    expires: t
+      .timestamp({
+        mode: "date",
+        withTimezone: true,
+      })
+      .notNull(),
   },
-  (session) => ({
-    userIdIdx: index("session_userId_idx").on(session.userId),
-  }),
+  (table) => [t.index("session_userId_idx").on(table.userId)],
 );
 
 export const sessionRelations = relations(sessions, ({ one }) => ({